assembler.py: Support sideband uniform (unifa)
It turns out V3D 4.x has a sideband uniform stream that is independent
from the ordinary one.  By writing an address to the "unifa" magic
register, which was named as "tmu" in V3D 3.x, one can read the content
of the memory with a ldunifa or ldunifarf signal.

See https://gitlab.freedesktop.org/mesa/mesa/-/merge_requests/8980 .